capsules: added readme and renamed process_info_driver to process_info
